#137 [refactor] 서비스 폴더 분리
