+++ src/services/auth/AuthService.ts
@@ -1,10 +1,10 @@
 import bcrypt from 'bcryptjs';
-import errorGenerator from '../errors/errorGenerator';
-import { LoginDto } from '../interfaces/auth/LoginDto';
-import { PostBaseResponseDto } from '../interfaces/common/PostBaseResponseDto';
-import User from '../models/User';
-import message from '../modules/responseMessage';
-import statusCode from '../modules/statusCode';
+import errorGenerator from '../../errors/errorGenerator';
+import { LoginDto } from '../../interfaces/auth/LoginDto';
+import { PostBaseResponseDto } from '../../interfaces/common/PostBaseResponseDto';
+import User from '../../models/User';
+import message from '../../modules/responseMessage';
+import statusCode from '../../modules/statusCode';
 
 const login = async (loginDto: LoginDto): Promise<PostBaseResponseDto> => {
   try {
