+++ src/services/index.ts
@@ -1,8 +1,20 @@
-import AuthService from './AuthService';
-import EventService from './EventService';
-import RoomService from './RoomService';
-import RuleService from './RuleService';
-import UserService from './UserService';
+import AuthService from './auth/AuthService';
+import EventRetrieveService from './event/EventRetrieveService';
+import EventService from './event/EventService';
+import RoomRetrieveService from './room/RoomRetrieveService';
+import RoomService from './room/RoomService';
+import RuleRetrieveService from './rule/RuleRetrieveService';
+import RuleService from './rule/RuleService';
+import UserService from './user/UserService';
 
 // service index file
-export { AuthService, UserService, RoomService, RuleService, EventService };
+export {
+  AuthService,
+  UserService,
+  RoomService,
+  RoomRetrieveService,
+  RuleService,
+  RuleRetrieveService,
+  EventService,
+  EventRetrieveService
+};
