+++ src/services/user/UserServiceUtils.ts
@@ -1,7 +1,7 @@
-import errorGenerator from '../errors/errorGenerator';
-import User from '../models/User';
-import message from '../modules/responseMessage';
-import statusCode from '../modules/statusCode';
+import errorGenerator from '../../errors/errorGenerator';
+import User from '../../models/User';
+import message from '../../modules/responseMessage';
+import statusCode from '../../modules/statusCode';
 
 const findUserById = async (userId: string) => {
   const user = await User.findById(userId);
